requirements.txt
jupyter==1.1.1
streamlit==1.46.1
seaborn==0.13.2
plotly-express==0.4.1
scikit-learn==1.7.0
shap==0.48.0
watchdog==6.0.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scipy==1.11.3
scanpy==1.9.5